chore: validate unknowns with zod
